// backend/src/site-config/dto/update-site-config.dto.ts
export class UpdateSiteConfigDto {
    appName?: string;
    seoTitle?: string;
    seoDescription?: string;
    heroTitle?: string;
    heroSubtitle?: string;
    heroBackgroundImage?: string;
    contactPhone?: string;
    contactEmail?: string;
    contactWhatsApp?: string;
    primaryColor?: string;
    layoutType?: string;
    showPrices?: boolean;
    showUnavailable?: boolean;
}